Start of adding some new functions (sleep, getCssVariable, setCssVariable). Better segregation of browser-related testing before expanding that testing. substr forever! Some documentation clean up.
